Add error code representing application error                                                                            ✘ 130 gql-error-application ✱
Take the update direct debit/payment schedule payment amount mutation
which uses the use case
```py
        (can_update, _, _,) = update_schedules.update_account_directdebit_amount(
            account=account,
            payment_amount=amount,
            preferred_valid_from=localtime.today(),
            account_user=user,
        )
        if not can_update:
            // return error here
```
It returns a can_update boolean which we want to use to notify the
API caller that the action was denied.

This is an error that stems from business rules not being met.
Therefore it feels appropriate to have an `errorClass` that represents
actions being denied due to application/business logic errors